requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
shap>=0.44.0
matplotlib>=3.7.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0